+++ src/main/java/de/michab/scream/binding/SchemeObject.java
@@ -16,8 +16,6 @@ import java.util.logging.Logger;
 
 import org.smack.util.JavaUtil;
 
-import de.michab.scream.Continuation;
-import de.michab.scream.Continuation.Thunk;
 import de.michab.scream.ConversionFailedX;
 import de.michab.scream.RuntimeX;
 import de.michab.scream.Scream.Scc;
@@ -35,7 +33,9 @@ import de.michab.scream.fcos.Symbol;
 import de.michab.scream.fcos.Syntax;
 import de.michab.scream.fcos.Vector;
 import de.michab.scream.pops.Primitives;
+import de.michab.scream.util.Continuation;
 import de.michab.scream.util.Scut;
+import de.michab.scream.util.Continuation.Thunk;
 
 /**
  * An instance of this class boxes an entity from the Java object system,
