+++ src/main/java/de/michab/scream/fcos/Cons.java
@@ -10,11 +10,11 @@ import java.util.HashSet;
 
 import org.smack.util.Holder;
 
-import de.michab.scream.Continuation.Thunk;
 import de.michab.scream.RuntimeX;
 import de.michab.scream.Scream.Scc;
 import de.michab.scream.fcos.Lambda.L;
 import de.michab.scream.util.Scut;
+import de.michab.scream.util.Continuation.Thunk;
 
 /**
  * Represents a list cell.  A list cell consists of two references called car
