+++ src/main/java/de/michab/scream/fcos/Number.java
@@ -5,11 +5,11 @@
  */
 package de.michab.scream.fcos;
 
-import de.michab.scream.Continuation.Thunk;
 import de.michab.scream.RuntimeX;
 import de.michab.scream.Scream.Scc;
 import de.michab.scream.pops.Primitives;
 import de.michab.scream.util.Scut;
+import de.michab.scream.util.Continuation.Thunk;
 
 /**
  * The base class for Scream's numeric types.
