+++ src/main/java/de/michab/scream/fcos/Syntax.java
@@ -5,9 +5,9 @@
  */
 package de.michab.scream.fcos;
 
-import de.michab.scream.Continuation.Thunk;
 import de.michab.scream.RuntimeX;
 import de.michab.scream.Scream.Scc;
+import de.michab.scream.util.Continuation.Thunk;
 
 /**
  * Followed the scheme spec in naming this class.  An alternate name would be
