+++ src/main/java/de/michab/scream/pops/SyntaxCase.java
@@ -7,7 +7,6 @@ package de.michab.scream.pops;
 
 import java.util.HashSet;
 
-import de.michab.scream.Continuation.Thunk;
 import de.michab.scream.RuntimeX;
 import de.michab.scream.Scream.Scc;
 import de.michab.scream.fcos.Cons;
@@ -17,6 +16,7 @@ import de.michab.scream.fcos.SchemeBoolean;
 import de.michab.scream.fcos.Symbol;
 import de.michab.scream.fcos.Syntax;
 import de.michab.scream.util.Scut;
+import de.michab.scream.util.Continuation.Thunk;
 
 /**
  * (case <key> <clause1> <clause2> ...) syntax; r7rs 15
