+++ src/main/java/de/michab/scream/pops/SyntaxLet.java
@@ -5,7 +5,6 @@
  */
 package de.michab.scream.pops;
 
-import de.michab.scream.Continuation.Thunk;
 import de.michab.scream.RuntimeX;
 import de.michab.scream.Scream.Scc;
 import de.michab.scream.fcos.Cons;
@@ -14,6 +13,7 @@ import de.michab.scream.fcos.FirstClassObject;
 import de.michab.scream.fcos.Symbol;
 import de.michab.scream.fcos.Syntax;
 import de.michab.scream.util.Scut;
+import de.michab.scream.util.Continuation.Thunk;
 import de.michab.scream.util.Scut.ConsumerX;
 
 /**
