+++ src/main/java/de/michab/scream/pops/SyntaxTime.java
@@ -7,7 +7,6 @@ package de.michab.scream.pops;
 
 import org.smack.util.TimeProbe;
 
-import de.michab.scream.Continuation.Thunk;
 import de.michab.scream.RuntimeX;
 import de.michab.scream.Scream.Scc;
 import de.michab.scream.fcos.Cons;
@@ -15,6 +14,7 @@ import de.michab.scream.fcos.Environment;
 import de.michab.scream.fcos.FirstClassObject;
 import de.michab.scream.fcos.SchemeInteger;
 import de.michab.scream.fcos.Syntax;
+import de.michab.scream.util.Continuation.Thunk;
 
 /**
  * (%time expression)
